flex: shrink the public API surface and move test-only code out of the library
Follow-up to the flex/ split, addressing the same review theme (keep the
interface module about the interface).

Public surface
- Restore __all__ in flex_attention_interface.py (it was dropped by the
  splitter) and cut it from 11 names to 5: the four entry points plus
  SUPPORT_STATUS. The backend-override registry (choose_backend,
  register_backend_override, clear_backend_overrides) and
  clear_classification_cache are internals used by tests and tuners; they stay
  importable from flex._routing / flex._cache but are no longer advertised.
  primus_turbo.pytorch.ops.attention already re-exported only the four entry
  points, so no import path changes.

Test-only code out of the library
- check_alibi_sign_convention, assert_alibi_sign_convention and
  _reference_attention_with_alibi (a dense fp32 reference) move verbatim from
  flex/_score_mod.py to tests/pytorch/ops/test_flex_score_mod.py, along with
  the _ASSUMED_ALIBI_SIGN constant from flex/_config.py. Nothing in
  primus_turbo called them.

Interface module holds signatures and control flow only
- The 98-line SUPPORT_STATUS manifest moves to flex/_support_status.py and is
  re-exported.

Verified: three moved functions AST-identical to the originals, 205 tests
still collected across the six files, flex/ dependency graph still acyclic,
pre-commit green.
