+++ primus_turbo/pytorch/ops/attention/flex/_score_mod.py
@@ -18,120 +18,13 @@ empirically whether the backend's ``alibi_slopes`` follows flex's
 """
 
 import math
-from typing import Any, Callable, Dict, Optional, Tuple
+from typing import Callable, Optional, Tuple
 
 import torch
 
 from ._cache import _ALIBI_CACHE, _CACHE_MISS, _SOFTCAP_CACHE, _cache_get, _cache_put
-from ._config import _ALIBI_TOL, _ASSUMED_ALIBI_SIGN, _SOFTCAP_TOL
+from ._config import _ALIBI_TOL, _SOFTCAP_TOL
 from ._probe import _call_score_mod
-
-
-def _reference_attention_with_alibi(
-    q_bhsd: torch.Tensor,
-    k_bhsd: torch.Tensor,
-    v_bhsd: torch.Tensor,
-    slopes: torch.Tensor,
-    sign: float,
-    *,
-    causal: bool = True,
-) -> torch.Tensor:
-    """Dense fp32 reference for ``softmax(QK^T/sqrt(d) + sign*slope[h]*(kv-q)) V``."""
-    q = q_bhsd.float()
-    k = k_bhsd.float()
-    v = v_bhsd.float()
-    s_q, s_k = q.shape[-2], k.shape[-2]
-    scores = torch.matmul(q, k.transpose(-1, -2)) / math.sqrt(q.shape[-1])
-    q_idx = torch.arange(s_q, device=q.device).view(s_q, 1)
-    kv_idx = torch.arange(s_k, device=q.device).view(1, s_k)
-    scores = scores + sign * slopes.float().view(1, -1, 1, 1) * (kv_idx - q_idx).float()
-    if causal:
-        scores = scores.masked_fill(kv_idx > q_idx, float("-inf"))
-    return torch.matmul(torch.softmax(scores, dim=-1), v)
-
-
-def check_alibi_sign_convention(
-    *,
-    device: Any = None,
-    dtype: torch.dtype = torch.float16,
-    seqlen: int = 256,
-    num_heads: int = 4,
-    head_dim: int = 64,
-    seed: int = 0,
-) -> Dict[str, Any]:
-    """Empirically determine this build's ALiBi sign convention.
-
-    The compat layer maps a flex ``score_mod`` of the form ``score + slope*(kv-q)``
-    onto ``flash_attn_func(alibi_slopes=slope)``. Whether the kernel adds
-    ``+slope*(kv-q)`` or ``-slope*(kv-q)`` is a property of the *installed aiter build*,
-    not of this file -- get it wrong and ALiBi results are silently incorrect. This
-    helper runs one small attention through the real backend and scores it against both
-    fp32 references, so a new container/build can be validated in seconds instead of
-    trusting a comment.
-
-    Returns a dict with ``sign`` (``+1.0``/``-1.0``, or ``None`` if neither reference
-    matches), ``plus_err`` / ``minus_err`` (relative L2), ``matches_assumption`` and the
-    ``assumed_sign`` this layer is coded against. Requires a GPU (it calls the real
-    kernel); it raises whatever the backend raises if one is unavailable.
-    """
-    from primus_turbo.pytorch.ops.attention.flash_attn_interface import flash_attn_func
-
-    if device is None:
-        device = "cuda"
-    gen = torch.Generator(device="cpu").manual_seed(seed)
-    shape = (1, seqlen, num_heads, head_dim)  # bshd for the Turbo backend
-    q = torch.randn(shape, generator=gen, dtype=torch.float32).to(device=device, dtype=dtype)
-    k = torch.randn(shape, generator=gen, dtype=torch.float32).to(device=device, dtype=dtype)
-    v = torch.randn(shape, generator=gen, dtype=torch.float32).to(device=device, dtype=dtype)
-    # Distinct, clearly separated slopes so the two sign hypotheses cannot alias.
-    slopes = torch.tensor([2.0 ** -(i + 1) for i in range(num_heads)], dtype=torch.float32, device=device)
-
-    out = flash_attn_func(q, k, v, causal=True, alibi_slopes=slopes)
-    if isinstance(out, tuple):
-        out = out[0]
-    got = out.transpose(1, 2).float()  # -> bhsd for comparison
-
-    q_b, k_b, v_b = (t.transpose(1, 2) for t in (q, k, v))
-
-    def _rel_l2(sign: float) -> float:
-        ref = _reference_attention_with_alibi(q_b, k_b, v_b, slopes, sign)
-        return float((got - ref).norm() / ref.norm().clamp_min(1e-12))
-
-    plus_err = _rel_l2(1.0)
-    minus_err = _rel_l2(-1.0)
-    # A match must be both small in absolute terms and decisively better than the other
-    # hypothesis; otherwise report "unknown" rather than guessing.
-    best_err, best_sign = min((plus_err, 1.0), (minus_err, -1.0))
-    other_err = max(plus_err, minus_err)
-    sign = best_sign if (best_err < 2e-2 and other_err > 10.0 * best_err) else None
-    return {
-        "sign": sign,
-        "plus_err": plus_err,
-        "minus_err": minus_err,
-        "assumed_sign": _ASSUMED_ALIBI_SIGN,
-        "matches_assumption": sign == _ASSUMED_ALIBI_SIGN,
-        "dtype": str(dtype),
-        "shape": {"seqlen": seqlen, "num_heads": num_heads, "head_dim": head_dim},
-    }
-
-
-def assert_alibi_sign_convention(**kwargs: Any) -> Dict[str, Any]:
-    """Run :func:`check_alibi_sign_convention` and raise unless it matches the assumption.
-
-    Intended as a one-line build gate (CI job, container smoke test, or the first thing
-    a new environment runs) so a flipped-sign build fails loudly instead of quietly
-    producing wrong ALiBi outputs.
-    """
-    report = check_alibi_sign_convention(**kwargs)
-    if not report["matches_assumption"]:
-        raise RuntimeError(
-            "Turbo flex compat layer: the ALiBi sign convention of this build does not match the "
-            f"assumed +slope*(kv-q) (assumed_sign={report['assumed_sign']}, measured "
-            f"sign={report['sign']}, plus_err={report['plus_err']:.3g}, "
-            f"minus_err={report['minus_err']:.3g}). ALiBi results would be silently wrong on this "
-            "build."
-        )
-    return report
 
 
 def _alibi_sample_points(q_len: int, kv_len: int) -> Tuple[Tuple[int, int], ...]:
